Fix e2e, move generate filters query, handle spacing
