+++ web/src/e2e/e2e.test.ts
@@ -1651,7 +1651,7 @@ describe('e2e test suite', () => {
             // Search for repo:gorilla in the repo filter chip input
             await driver.page.keyboard.type('gorilla')
             await driver.page.keyboard.press('Enter')
-            await driver.assertWindowLocation('/search?repo=gorilla&q=&patternType=literal')
+            await driver.assertWindowLocation('/search?q=repo:gorilla&patternType=literal')
 
             // Edit the filter
             await driver.page.waitForSelector('.filter-input')
@@ -1661,7 +1661,7 @@ describe('e2e test suite', () => {
             // Press enter to lock in filter
             await driver.page.keyboard.press('Enter')
             // The main query input should be autofocused, so hit enter again to submit
-            await driver.assertWindowLocation('/search?repo=gorilla/mux&q=&patternType=literal')
+            await driver.assertWindowLocation('/search?q=repo:gorilla/mux&patternType=literal')
 
             // Add a file filter from search results page
             await driver.page.waitForSelector('.e2e-add-filter-button-file', { visible: true })
@@ -1670,7 +1670,7 @@ describe('e2e test suite', () => {
             await driver.page.keyboard.type('README')
             await driver.page.keyboard.press('Enter')
             await driver.page.keyboard.press('Enter')
-            await driver.assertWindowLocation('/search?repo=gorilla/mux&file=README&q=&patternType=literal')
+            await driver.assertWindowLocation('/search?q=repo:gorilla/mux+file:README&patternType=literal')
 
             // Delete filter
             await driver.page.goto(sourcegraphBaseUrl + '/search?repo=gorilla/mux&q=&patternType=literal')
@@ -1719,7 +1719,7 @@ describe('e2e test suite', () => {
             await driver.page.waitForSelector('.e2e-filter-input-radio-button-no')
             await driver.page.click('.e2e-filter-input-radio-button-no')
             await driver.page.click('.e2e-confirm-filter-button')
-            await driver.assertWindowLocation('/search?fork=no&q=test&patternType=literal')
+            await driver.assertWindowLocation('/search?q=test+fork:no&patternType=literal')
             // Edit filter
             await driver.page.waitForSelector('.filter-input')
             await driver.page.waitForSelector('.e2e-filter-input__button-text-fork')
